scrapper/requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
tqdm==4.62.3
Pillow==8.4.0
colorama==0.4.6